Bug 1820773 - Refactor Settings test classes to reflect current categories
