+++ fenix/app/src/androidTest/java/org/mozilla/fenix/ui/SettingsDeleteBrowsingDataTest.kt
@@ -1,0 +1,256 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+package org.mozilla.fenix.ui
+
+import okhttp3.mockwebserver.MockWebServer
+import org.junit.After
+import org.junit.Before
+import org.junit.Rule
+import org.junit.Test
+import org.mozilla.fenix.R
+import org.mozilla.fenix.customannotations.SmokeTest
+import org.mozilla.fenix.helpers.AndroidAssetDispatcher
+import org.mozilla.fenix.helpers.HomeActivityIntentTestRule
+import org.mozilla.fenix.helpers.TestAssetHelper
+import org.mozilla.fenix.helpers.TestAssetHelper.getStorageTestAsset
+import org.mozilla.fenix.helpers.TestHelper.exitMenu
+import org.mozilla.fenix.helpers.TestHelper.getStringResource
+import org.mozilla.fenix.helpers.TestHelper.mDevice
+import org.mozilla.fenix.helpers.TestHelper.restartApp
+import org.mozilla.fenix.helpers.TestHelper.setNetworkEnabled
+import org.mozilla.fenix.ui.robots.browserScreen
+import org.mozilla.fenix.ui.robots.homeScreen
+import org.mozilla.fenix.ui.robots.navigationToolbar
+import org.mozilla.fenix.ui.robots.settingsScreen
+
+/**
+ *  Tests for verifying the Settings for:
+ *  Delete Browsing Data
+ *  Delete Browsing Data on quit
+ *
+ */
+
+class SettingsDeleteBrowsingDataTest {
+    /* ktlint-disable no-blank-line-before-rbrace */ // This imposes unreadable grouping.
+    private lateinit var mockWebServer: MockWebServer
+
+    @get:Rule
+    val activityTestRule = HomeActivityIntentTestRule.withDefaultSettingsOverrides(skipOnboarding = true)
+
+    @Before
+    fun setUp() {
+        mockWebServer = MockWebServer().apply {
+            dispatcher = AndroidAssetDispatcher()
+            start()
+        }
+    }
+
+    @After
+    fun tearDown() {
+        mockWebServer.shutdown()
+    }
+
+    @Test
+    fun deleteBrowsingDataOptionStatesTest() {
+        homeScreen {
+        }.openThreeDotMenu {
+        }.openSettings {
+        }.openSettingsSubMenuDeleteBrowsingData {
+            verifyAllCheckBoxesAreChecked()
+            switchBrowsingHistoryCheckBox()
+            switchCachedFilesCheckBox()
+            verifyOpenTabsCheckBox(true)
+            verifyBrowsingHistoryDetails(false)
+            verifyCookiesCheckBox(true)
+            verifyCachedFilesCheckBox(false)
+            verifySitePermissionsCheckBox(true)
+            verifyDownloadsCheckBox(true)
+        }
+
+        restartApp(activityTestRule)
+
+        homeScreen {
+        }.openThreeDotMenu {
+        }.openSettings {
+        }.openSettingsSubMenuDeleteBrowsingData {
+            verifyOpenTabsCheckBox(true)
+            verifyBrowsingHistoryDetails(false)
+            verifyCookiesCheckBox(true)
+            verifyCachedFilesCheckBox(false)
+            verifySitePermissionsCheckBox(true)
+            verifyDownloadsCheckBox(true)
+            switchOpenTabsCheckBox()
+            switchBrowsingHistoryCheckBox()
+            switchCookiesCheckBox()
+            switchCachedFilesCheckBox()
+            switchSitePermissionsCheckBox()
+            switchDownloadsCheckBox()
+            verifyOpenTabsCheckBox(false)
+            verifyBrowsingHistoryDetails(true)
+            verifyCookiesCheckBox(false)
+            verifyCachedFilesCheckBox(true)
+            verifySitePermissionsCheckBox(false)
+            verifyDownloadsCheckBox(false)
+        }
+
+        restartApp(activityTestRule)
+
+        homeScreen {
+        }.openThreeDotMenu {
+        }.openSettings {
+        }.openSettingsSubMenuDeleteBrowsingData {
+            verifyOpenTabsCheckBox(false)
+            verifyBrowsingHistoryDetails(true)
+            verifyCookiesCheckBox(false)
+            verifyCachedFilesCheckBox(true)
+            verifySitePermissionsCheckBox(false)
+            verifyDownloadsCheckBox(false)
+        }
+    }
+
+    @Test
+    fun deleteTabsDataWithNoOpenTabsTest() {
+        homeScreen {
+        }.openThreeDotMenu {
+        }.openSettings {
+        }.openSettingsSubMenuDeleteBrowsingData {
+            verifyAllCheckBoxesAreChecked()
+            selectOnlyOpenTabsCheckBox()
+            clickDeleteBrowsingDataButton()
+            confirmDeletionAndAssertSnackbar()
+        }
+        settingsScreen {
+            verifyGeneralHeading()
+        }
+    }
+
+    @SmokeTest
+    @Test
+    fun deleteTabsDataTest() {
+        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
+
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
+            mDevice.waitForIdle()
+        }.openThreeDotMenu {
+        }.openSettings {
+        }.openSettingsSubMenuDeleteBrowsingData {
+            verifyAllCheckBoxesAreChecked()
+            selectOnlyOpenTabsCheckBox()
+            clickDeleteBrowsingDataButton()
+            clickDialogCancelButton()
+            verifyOpenTabsCheckBox(true)
+            clickDeleteBrowsingDataButton()
+            confirmDeletionAndAssertSnackbar()
+        }
+        settingsScreen {
+            verifyGeneralHeading()
+        }.openSettingsSubMenuDeleteBrowsingData {
+            verifyOpenTabsDetails("0")
+        }.goBack {
+        }.goBack {
+        }.openTabDrawer {
+            verifyNoOpenTabsInNormalBrowsing()
+        }
+    }
+
+    @SmokeTest
+    @Test
+    fun deleteBrowsingHistoryAndSiteDataTest() {
+        val storageWritePage = getStorageTestAsset(mockWebServer, "storage_write.html").url
+        val storageCheckPage = getStorageTestAsset(mockWebServer, "storage_check.html").url
+
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(storageWritePage) {
+        }.openNavigationToolbar {
+        }.enterURLAndEnterToBrowser(storageCheckPage) {
+            verifyPageContent("Session storage has value")
+            verifyPageContent("Local storage has value")
+        }.openThreeDotMenu {
+        }.openSettings {
+        }.openSettingsSubMenuDeleteBrowsingData {
+            verifyBrowsingHistoryDetails("2")
+            selectOnlyBrowsingHistoryCheckBox()
+            clickDeleteBrowsingDataButton()
+            clickDialogCancelButton()
+            verifyBrowsingHistoryDetails(true)
+            clickDeleteBrowsingDataButton()
+            confirmDeletionAndAssertSnackbar()
+            verifyBrowsingHistoryDetails("0")
+            exitMenu()
+        }
+        navigationToolbar {
+        }.openThreeDotMenu {
+        }.openHistory {
+            verifyEmptyHistoryView()
+            mDevice.pressBack()
+        }
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(storageCheckPage) {
+            verifyPageContent("Session storage empty")
+            verifyPageContent("Local storage empty")
+        }
+    }
+
+    @SmokeTest
+    @Test
+    fun deleteCookiesTest() {
+        val genericPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
+        val cookiesTestPage = getStorageTestAsset(mockWebServer, "storage_write.html").url
+
+        // Browsing a generic page to allow GV to load on a fresh run
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(genericPage.url) {
+        }.openNavigationToolbar {
+        }.enterURLAndEnterToBrowser(cookiesTestPage) {
+            verifyPageContent("No cookies set")
+            clickSetCookiesButton()
+            verifyPageContent("user=android")
+        }.openThreeDotMenu {
+        }.openSettings {
+        }.openSettingsSubMenuDeleteBrowsingData {
+            selectOnlyCookiesCheckBox()
+            clickDeleteBrowsingDataButton()
+            confirmDeletionAndAssertSnackbar()
+            exitMenu()
+        }
+        browserScreen {
+        }.openThreeDotMenu {
+        }.refreshPage {
+            verifyPageContent("No cookies set")
+        }
+    }
+
+    @SmokeTest
+    @Test
+    fun deleteCachedFilesTest() {
+        val pocketTopArticles = getStringResource(R.string.pocket_pinned_top_articles)
+
+        homeScreen {
+            verifyExistingTopSitesTabs(pocketTopArticles)
+        }.openTopSiteTabWithTitle(pocketTopArticles) {
+            waitForPageToLoad()
+        }.openTabDrawer {
+        }.openNewTab {
+        }.submitQuery("about:cache") {
+            // disabling wifi to prevent downloads in the background
+            setNetworkEnabled(enabled = false)
+        }.openThreeDotMenu {
+        }.openSettings {
+        }.openSettingsSubMenuDeleteBrowsingData {
+            selectOnlyCachedFilesCheckBox()
+            clickDeleteBrowsingDataButton()
+            confirmDeletionAndAssertSnackbar()
+            exitMenu()
+        }
+        browserScreen {
+        }.openThreeDotMenu {
+        }.refreshPage {
+            verifyNetworkCacheIsEmpty("memory")
+            verifyNetworkCacheIsEmpty("disk")
+        }
+        setNetworkEnabled(enabled = true)
+    }
+}
