+++ fenix/app/src/androidTest/java/org/mozilla/fenix/ui/SettingsPrivateBrowsingTest.kt
@@ -1,0 +1,190 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+package org.mozilla.fenix.ui
+
+import okhttp3.mockwebserver.MockWebServer
+import org.junit.After
+import org.junit.Before
+import org.junit.Rule
+import org.junit.Test
+import org.mozilla.fenix.helpers.AndroidAssetDispatcher
+import org.mozilla.fenix.helpers.HomeActivityIntentTestRule
+import org.mozilla.fenix.helpers.TestAssetHelper
+import org.mozilla.fenix.helpers.TestHelper
+import org.mozilla.fenix.helpers.TestHelper.mDevice
+import org.mozilla.fenix.ui.robots.addToHomeScreen
+import org.mozilla.fenix.ui.robots.browserScreen
+import org.mozilla.fenix.ui.robots.homeScreen
+import org.mozilla.fenix.ui.robots.navigationToolbar
+
+class SettingsPrivateBrowsingTest {
+    private lateinit var mockWebServer: MockWebServer
+    private val pageShortcutName = TestHelper.generateRandomString(5)
+
+    @get:Rule
+    val activityTestRule = HomeActivityIntentTestRule.withDefaultSettingsOverrides(skipOnboarding = true)
+
+    @Before
+    fun setUp() {
+        mockWebServer = MockWebServer().apply {
+            dispatcher = AndroidAssetDispatcher()
+            start()
+        }
+    }
+
+    @After
+    fun tearDown() {
+        mockWebServer.shutdown()
+    }
+
+    @Test
+    fun verifyPrivateBrowsingMenuItemsTest() {
+        homeScreen {
+        }.openThreeDotMenu {
+        }.openSettings {
+        }.openPrivateBrowsingSubMenu {
+            verifyAddPrivateBrowsingShortcutButton()
+            verifyOpenLinksInPrivateTab()
+            verifyOpenLinksInPrivateTabOff()
+        }.goBack {
+            verifySettingsView()
+        }
+    }
+
+    @Test
+    fun openExternalLinksInPrivateTest() {
+        val firstWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
+        val secondWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 2)
+
+        setOpenLinksInPrivateOn()
+
+        TestHelper.openAppFromExternalLink(firstWebPage.url.toString())
+
+        browserScreen {
+            verifyUrl(firstWebPage.url.toString())
+        }.openTabDrawer {
+            verifyPrivateModeSelected()
+        }.closeTabDrawer {
+        }.goToHomescreen { }
+
+        setOpenLinksInPrivateOff()
+
+        // We need to open a different link, otherwise it will open the same session
+        TestHelper.openAppFromExternalLink(secondWebPage.url.toString())
+
+        browserScreen {
+            verifyUrl(secondWebPage.url.toString())
+        }.openTabDrawer {
+            verifyNormalModeSelected()
+        }
+    }
+
+    @Test
+    fun launchPageShortcutInPrivateModeTest() {
+        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
+
+        setOpenLinksInPrivateOn()
+
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
+        }.openThreeDotMenu {
+        }.openAddToHomeScreen {
+            addShortcutName(pageShortcutName)
+            clickAddShortcutButton()
+            clickAddAutomaticallyButton()
+            verifyShortcutAdded(pageShortcutName)
+        }
+
+        mDevice.waitForIdle()
+        // We need to close the existing tab here, to open a different session
+        TestHelper.restartApp(activityTestRule)
+        browserScreen {
+        }.openTabDrawer {
+            closeTab()
+        }
+
+        addToHomeScreen {
+        }.searchAndOpenHomeScreenShortcut(pageShortcutName) {
+        }.openTabDrawer {
+            verifyPrivateModeSelected()
+        }
+    }
+
+    @Test
+    fun launchLinksInPrivateToggleOffStateDoesntChangeTest() {
+        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
+
+        setOpenLinksInPrivateOn()
+
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
+        }.openThreeDotMenu {
+            expandMenu()
+        }.openAddToHomeScreen {
+            addShortcutName(pageShortcutName)
+            clickAddShortcutButton()
+            clickAddAutomaticallyButton()
+        }.openHomeScreenShortcut(pageShortcutName) {
+        }.goToHomescreen { }
+
+        setOpenLinksInPrivateOff()
+        TestHelper.restartApp(activityTestRule)
+        mDevice.waitForIdle()
+
+        addToHomeScreen {
+        }.searchAndOpenHomeScreenShortcut(pageShortcutName) {
+        }.openTabDrawer {
+            verifyNormalModeSelected()
+        }.closeTabDrawer {
+        }.openThreeDotMenu {
+        }.openSettings {
+        }.openPrivateBrowsingSubMenu {
+            verifyOpenLinksInPrivateTabOff()
+        }
+    }
+
+    @Test
+    fun addPrivateBrowsingShortcut() {
+        homeScreen {
+        }.openThreeDotMenu {
+        }.openSettings {
+        }.openPrivateBrowsingSubMenu {
+            cancelPrivateShortcutAddition()
+            addPrivateShortcutToHomescreen()
+            verifyPrivateBrowsingShortcutIcon()
+        }.openPrivateBrowsingShortcut {
+            verifySearchView()
+        }.openBrowser {
+        }.openTabDrawer {
+            verifyPrivateModeSelected()
+        }
+    }
+}
+
+private fun setOpenLinksInPrivateOn() {
+    homeScreen {
+    }.openThreeDotMenu {
+    }.openSettings {
+    }.openPrivateBrowsingSubMenu {
+        verifyOpenLinksInPrivateTabEnabled()
+        clickOpenLinksInPrivateTabSwitch()
+    }.goBack {
+    }.goBack {
+        verifyHomeComponent()
+    }
+}
+
+private fun setOpenLinksInPrivateOff() {
+    homeScreen {
+    }.openThreeDotMenu {
+    }.openSettings {
+    }.openPrivateBrowsingSubMenu {
+        clickOpenLinksInPrivateTabSwitch()
+        verifyOpenLinksInPrivateTabOff()
+    }.goBack {
+    }.goBack {
+        verifyHomeComponent()
+    }
+}
