+++ fenix/app/src/androidTest/java/org/mozilla/fenix/ui/SettingsSitePermissionsTest.kt
@@ -1,0 +1,52 @@
+package org.mozilla.fenix.ui
+
+import androidx.test.filters.SdkSuppress
+import org.junit.Rule
+import org.junit.Test
+import org.mozilla.fenix.customannotations.SmokeTest
+import org.mozilla.fenix.helpers.HomeActivityTestRule
+import org.mozilla.fenix.ui.robots.homeScreen
+
+class SettingsSitePermissionsTest {
+    @get:Rule
+    val activityTestRule = HomeActivityTestRule.withDefaultSettingsOverrides()
+
+    // Verifies that you can go to System settings and change app's permissions from inside the app
+    @SmokeTest
+    @Test
+    @SdkSuppress(minSdkVersion = 29)
+    fun redirectToAppPermissionsSystemSettingsTest() {
+        homeScreen {
+        }.openThreeDotMenu {
+        }.openSettings {
+        }.openSettingsSubMenuSitePermissions {
+        }.openCamera {
+            verifyBlockedByAndroid()
+        }.goBack {
+        }.openLocation {
+            verifyBlockedByAndroid()
+        }.goBack {
+        }.openMicrophone {
+            verifyBlockedByAndroid()
+            clickGoToSettingsButton()
+            openAppSystemPermissionsSettings()
+            switchAppPermissionSystemSetting("Camera", "Allow")
+            goBackToSystemAppPermissionSettings()
+            verifySystemGrantedPermission("Camera")
+            switchAppPermissionSystemSetting("Location", "Allow")
+            goBackToSystemAppPermissionSettings()
+            verifySystemGrantedPermission("Location")
+            switchAppPermissionSystemSetting("Microphone", "Allow")
+            goBackToSystemAppPermissionSettings()
+            verifySystemGrantedPermission("Microphone")
+            goBackToPermissionsSettingsSubMenu()
+            verifyUnblockedByAndroid()
+        }.goBack {
+        }.openLocation {
+            verifyUnblockedByAndroid()
+        }.goBack {
+        }.openCamera {
+            verifyUnblockedByAndroid()
+        }
+    }
+}
